Fix a bug in DigestMD5SaslClient (authzid vs authcid)
The 'username' value for DIGEST-MD5 must not be authzid, which is
according to RFC6120 6.3.8 null or empty in most cases. Use m_name
aka. authcid for 'username' and add the authzid to the appropriate place
within the digest response as described in RFC2831 2.1.2.

This appears to be a bug in OpenLDAP. Reported as
http://www.openldap.org/its/index.cgi?findid=7817